+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Fields `name`, `nameComplete`, `ean` and `Videos` to SKU.
+### Fixed
+- Different SKU with the same image.
+
 ## [1.33.0] - 2021-02-01
 
 ### Added